refactor(tests): rename E2E test classes from Team to Club
- git mv Features/Teams/ → Features/Clubs/ (E2E); rename classes TeamsIndexTests→ClubsIndexTests, TeamDetailsTests→ClubDetailsTests, CreateTeamPageTests→CreateClubPageTests
- Preserve all GotoAsync("/teams/...") URL strings and href strings unchanged (routes stay /teams)
